Remove default exports from stories.
Storybook was spewing warnings about using default exports in the
same file as storiesOf(), so I modified the stories to remove the
default keyword. The exports are still available, but they're no
longer the default. The unit tests were modified to use the new
non-default exports from the stories. In a couple of cases,
inconsistencies in the names of the imported/exported items were
fixed.